+++ lib/src/certs/unregister_drep_cert.c
@@ -134,7 +134,7 @@ cardano_unregister_drep_cert_from_cbor(cardano_cbor_reader_t* reader, cardano_un
 
   static const char* validator_name = "unregister_drep_cert";
 
-  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, EMBEDDED_GROUP_SIZE);
+  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
