+++ lib/src/common/credential.c
@@ -224,7 +224,7 @@ cardano_credential_from_cbor(cardano_cbor_reader_t* reader, cardano_credential_t
 
   static const char* validator_name = "Credential";
 
-  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, CREDENTIAL_ARRAY_SIZE);
+  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)CREDENTIAL_ARRAY_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
