+++ lib/src/common/protocol_version.c
@@ -116,7 +116,7 @@ cardano_protocol_version_from_cbor(cardano_cbor_reader_t* reader, cardano_protoc
 
   static const char* validator_name = "protocol_version";
 
-  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, PROTOCOL_VERSION_EMBEDDED_GROUP_SIZE);
+  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)PROTOCOL_VERSION_EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
