+++ lib/src/pool_params/multi_host_name_relay.c
@@ -128,7 +128,7 @@ cardano_multi_host_name_relay_from_cbor(cardano_cbor_reader_t* reader, cardano_m
 
   static const char* validator_name = "multi_host_name_relay";
 
-  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, EMBEDDED_GROUP_SIZE);
+  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
